Moved files to follow the same structure as master
