+++ back-end-reference/package.json
@@ -8,6 +8,6 @@
     "express": "^4.17.1"
   },
   "scripts": {
-    "start": "node server.js"
+    "start": "node index.js"
   }
 }
