+++ back-end/app.js
@@ -1,6 +1,10 @@
 const express = require('express')
 const app = express()
+
 const multipleStore = require("./multiStore");
+const stampTotalTrend = require("./stampTotalTrend");
+const totalRedeemedTotalUnredeemed = require("./totalRedeemedTotalUnredeemed");
+const actualTotalsPie = require("./actualTotalsPie");
 
 const port = 3000;
 
@@ -16,6 +20,21 @@ app.get('/', (req, res) => {
 app.get('/multi_store_customers', (req, res) => {
   enableLocalCors(res)
   multipleStore(req, res)
+})
+
+app.get('/stamp_total_trend', (req, res) => {
+  enableLocalCors(res)
+  stampTotalTrend(req, res)
+})
+
+app.get('/total_redeemed_total_unredeemed', (req, res) => {
+  enableLocalCors(res)
+  totalRedeemedTotalUnredeemed(req, res)
+})
+
+app.get('/actual_totals_pie', (req, res) => {
+  enableLocalCors(res)
+  actualTotalsPie(req, res)
 })
 
 app.listen(port, () => {
